# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Cria um novo projeto de identidade visual",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos no payload"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "Lista os projetos do usuário autenticado",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Texto para filtrar projetos por nome ou descrição",
            "schema": {
              "example": "campanha",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Quantidade de itens por página",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de projetos retornada com sucesso"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Busca um projeto pelo identificador",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "summary": "Atualiza os dados de um projeto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ID ou payload inválido"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "summary": "Remove um projeto existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "ID inválido"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/projects/{projectId}/palettes": {
      "post": {
        "operationId": "PaletteController_create",
        "summary": "Cria paleta vinculada a um projeto",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaletteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Paleta criada com sucesso"
          },
          "401": {
            "description": "Token JWT ausente ou inválido"
          }
        },
        "tags": [
          "palettes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PaletteController_findAll",
        "summary": "Lista paletas de um projeto",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paletas retornadas com sucesso"
          }
        },
        "tags": [
          "palettes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/projects/{projectId}/palettes/{id}": {
      "get": {
        "operationId": "PaletteController_findOne",
        "summary": "Busca uma paleta pelo identificador",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da paleta",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "UUID do projeto",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Paleta encontrada com sucesso"
          }
        },
        "tags": [
          "palettes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PaletteController_update",
        "summary": "Atualiza dados de uma paleta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da paleta",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaletteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paleta atualizada com sucesso"
          }
        },
        "tags": [
          "palettes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PaletteController_remove",
        "summary": "Remove uma paleta existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID da paleta",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "palettes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Letzify API",
    "description": "Documentação da API Letzify",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Brandbook 2026"
          },
          "description": {
            "type": "string",
            "example": "Projeto de identidade visual da campanha 2026"
          },
          "status": {
            "type": "string",
            "enum": [
              "ALIGNMENT",
              "IN_PROGRESS",
              "FINISHED"
            ],
            "example": "ALIGNMENT"
          }
        },
        "required": [
          "name"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3f3a2c0f-92cc-4a49-a8c5-8e3de7f4f2c1"
          },
          "ownerId": {
            "type": "string",
            "example": "3f3a2c0f-92cc-4a49-a8c5-8e3de7f4f2c1"
          },
          "name": {
            "type": "string",
            "example": "Brandbook 2026"
          },
          "slug": {
            "type": "string",
            "example": "brandbook-2026"
          },
          "description": {
            "type": "string",
            "example": "Projeto de identidade visual da campanha 2026"
          },
          "status": {
            "type": "string",
            "enum": [
              "ALIGNMENT",
              "IN_PROGRESS",
              "FINISHED"
            ],
            "example": "ALIGNMENT"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-03T10:15:30.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-03T10:15:30.000Z"
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "slug",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePaletteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Paleta Principal"
          },
          "description": {
            "type": "string",
            "example": "Paleta principal do projeto"
          },
          "isPrimary": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdatePaletteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Paleta Principal"
          },
          "description": {
            "type": "string",
            "example": "Paleta principal do projeto"
          },
          "isPrimary": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}